requirements.txt
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
colored==2.2.4
data-printer==0.0.8
et_xmlfile==2.0.0
h11==0.14.0
idna==3.10
MouseInfo==0.1.3
numpy==2.2.0
opencv-python==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.3
pillow==11.0.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.2
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
